requirements.txt
Flask==3.0.3;
Flask-SQLAlchemy==3.1.1;
Flask-Login==0.6.3;
Werkzeug==3.0.4;
python-dotenv==1.0.1;
pytest==8.3.3;